introduce commands
Printing all the packages binaries doesn't scale really well. When
importing a build dependency for example, we don't care about the
associated binary if there is one.

Commands are now used to declare all the commands available, with a help
menu. If the `alias` attribute is passed, the command is replaced by the
alias.