introduce language.go
This is our first language support!

When "language.go" is imported, it adds the go package to the
environment and sets GO111MODULE.

This changes the extra modules import rule a little bit. Now that other
prefixes than `extra.` is being used, the modules doc will have to be
split in two to remove confusion on what is part of the core modules and
what is part of the extra modules.