feat(devshell): support default "$PRJ_ROOT" value
via the new option "devshell.prj_root_fallback".  This option, if set to
a non-null value, is used to set the value of "$PRJ_ROOT" in case it is
not set in the controlling environment, the entrypoint script did not
get a `--prj-root` option, "$IN_NIX_SHELL" is empty or unset, and
"$DIRENV_IN_ENVRC" is unset or any value other than "1".

The default option definition sets "$PRJ_ROOT" to the value of "$PWD".